Add a startup activity for the status bar
Previously the Widget would need to be disabled/enabled after every restart due to the ExperimentalSettings flag not being available in memory when the StatusBarManager starts up.

This registers a startup activity that only runs after the project (including the application level settings) have loaded.

It also moves the settings watcher into this class to keep the Factory "pure".

Lastly, default the widget to display as soon as the experimental setting is enabled. I think the setting used to be at project level, but it is now at application level, and it makes sense to me that if I enable it in Experimental Settings, it should immediately appear in the status bar.
